Simpler routers, split off socket, almost back to using Proxy on the client
