--- examples/README.md
@@ -1,7 +1,0 @@
-# [Cav](https://cav.bar) / Examples
-
-Shortcut: https://cav.bar/examples
-
-- [Blog](./blog): Markdown blogging
-- [Chat](./chat): Ephemeral chat rooms
-- [Shortcuts](./shortcuts): URL Shortening
