--- examples/blog/README.md
@@ -1,5 +1,0 @@
-# [Cav](https://cav.bar) / [Examples](../README.md) / Blog
-
-Minimalist blogging with a single `server.ts` and some markdown files.
-
-Shortcut: https://cav.bar/examples/blog
