--- examples/blog/header.md
@@ -1,7 +1,0 @@
----
-title: hey
----
-
-# Hello
-
-world
