--- examples/chat/assets/auth.css
@@ -1,33 +1,0 @@
-/* Copyright 2022 Connor Logan. All rights reserved. MIT License. */
-
-body {
-  justify-content: center;
-}
-
-.form {
-  display: flex;
-  flex-direction: column;
-  color: white;
-  position: relative;
-}
-
-.name {
-  font-size: 2em;
-  border-radius: 0.5em;
-  align-self: stretch;
-}
-
-.submit {
-  margin: 1.5em 0 0 0;
-  font-weight: 700;
-  font-size: 1.25em;
-  background: transparent;
-  padding: 0;
-  position: absolute;
-  top: calc(50% + 3.5em);
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
-.submit:disabled {
-  visibility: hidden;
-}
