--- examples/chat/assets/base.css
@@ -1,57 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  text-rendering: optimizeLegibility;
-  letter-spacing: 0.01em;
-}
-
-:root {
-  width: 100%;
-  height: 100%;
-  font-size: 16px;
-  font-family: sans-serif;
-}
-
-body {
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-  width: 100%;
-  min-height: 100%;
-  background: #1E1E1E;
-  line-height: 1.1;
-  letter-spacing: 0.01em;
-  position: relative;
-}
-
-a {
-  text-decoration: none;
-  font: inherit;
-}
-
-button {
-  font: inherit;
-  background: black;
-  color: white;
-  border: 0;
-  border-top: 3px solid transparent;
-  border-radius: 0.5em;
-  padding: 0.5em 0.75em;
-  cursor: pointer;
-}
-button:disabled {
-  cursor: default;
-}
-
-input,
-textarea {
-  appearance: none;
-  background: #2b2b2d;
-  color: #eee;
-  padding: 0.55em 0.75em 0.5em 0.75em;
-  border-radius: 1rem;
-  font: inherit;
-  outline: none;
-  border: 0;
-}
