--- examples/chat/assets/chat.css
@@ -1,82 +1,0 @@
-/* Copyright 2022 Connor Logan. All rights reserved. MIT License. */
-
-.room {
-  width: 100%;
-  max-width: 35em;
-  flex: 1 0;
-  display: flex;
-  flex-direction: column;
-  padding: 1.5em 1.5em 0 1.5em;
-}
-
-.messages {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-end;
-}
-
-.input {
-  display: grid;
-  align-items: stretch;
-  position: relative;
-  margin: 2em 0;
-  border-radius: 1.5em;
-}
-.input::after {
-  content: attr(data-value) " ";
-  white-space: pre-wrap;
-  visibility: hidden;
-  grid-area: 1 / 1;
-  padding: 0.75em 1em;
-}
-.input textarea {
-  grid-area: 1 / 1;
-  resize: none;
-  border-radius: 1.5em;
-  padding: 0.75em 1em;
-  overflow: hidden;
-}
-.input textarea:focus {
-  border-color: transparent;
-}
-.input textarea:focus::placeholder {
-  color: transparent;
-}
-
-/* html.room.messageGroup() */
-
-.group {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  max-width: 60%;
-  margin: 0.5em 0;
-}
-.group.right {
-  align-items: flex-end;
-  align-self: flex-end;
-}
-
-.user {
-  color: #9A9A9A;
-  font-size: 0.9em;
-  margin: 0.25em 0;
-}
-
-/* html.room.message() */
-
-.msg {
-  background: #3B3B3D;
-  color: white;
-  padding: 0.6em 0.8em 0.5em 0.8em;
-  text-align: left;
-  margin: 0.15em 0;
-  border-radius: 1.25em;
-}
-.right .user + .msg {
-  border-radius: 1.25em 0.25em 1.25em 1.25em;
-}
-.user + .msg {
-  border-radius: 0.25em 1.25em 1.25em 1.25em;
-}
