--- examples/chat/assets/index.css
@@ -1,21 +1,0 @@
-body {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.chat {
-  display: block;
-  cursor: pointer;
-  background: #444;
-  color: #eee;
-  padding: 0.5em 0.75em;
-  border: 0;
-  border-top: 5px solid transparent; /* fix */
-  font-size: 2em;
-  border-radius: 1rem;
-}
-.chat.disabled {
-  cursor: default !important;
-  opacity: 0.75;
-}
