--- examples/chat/controller/rpc/deps.ts
@@ -1,4 +1,0 @@
-// Copyright 2022 Connor Logan. All rights reserved. MIT License.
-// This module is browser-compatible.
-
-export * from "../../../client.ts";
