--- examples/chat/controller/rpc/mod.ts
@@ -1,4 +1,0 @@
-// Copyright 2022 Connor Logan. All rights reserved. MIT License.
-// This module is browser-compatible.
-
-export * as room from "./room.ts";
