--- examples/chat/deno.json
@@ -1,6 +1,0 @@
-{
-  "tasks": {
-    "dev": "DEV=1 deno run --watch --allow-env --allow-net --allow-read server.ts",
-    "prod": "deno run --allow-net --allow-read server.ts"
-  }
-}
