--- examples/chat/deps_dom.ts
@@ -1,4 +1,0 @@
-// Copyright 2022 Connor Logan. All rights reserved. MIT License.
-// This module is browser-only.
-
-export * from "../../dom.ts";
