--- examples/chat/deps_iso.ts
@@ -1,5 +1,0 @@
-// Copyright 2022 Connor Logan. All rights reserved. MIT License.
-// This module is browser-compatible.
-
-export * from "../../client.ts";
-export * from "../../html.ts";
