--- examples/chat/model.ts
@@ -1,3 +1,0 @@
-// Copyright 2022 Connor Logan. All rights reserved. MIT License.
-
-export * as room from "./room/api.ts";
