+++ examples/chat/model/deps.ts
@@ -1,1 +1,3 @@
+// Copyright 2022 Connor Logan. All rights reserved. MIT License.
+
 export * from "../../../mod.ts";
