+++ examples/chat/view/dom/deps.ts
@@ -1,4 +1,4 @@
 // Copyright 2022 Connor Logan. All rights reserved. MIT License.
 // This module is browser-only.
 
-export * from "../../../dom.ts";
+export * from "../../../../dom.ts";
